Intellisense never auto-commits for F#, not even for '.'
Fixes vsts #563779 (#3798)

Determined that it is too difficult to detect all the occurrences of
identifiers in F# code for the time being, so it is hard to determine that the
popup should not be displayed. Given this, we should be far less aggressive
about auto-committing (even when "Complete with Space or Punctuation" is
switched on. This is a good default for C#, but a bad default for F#.)

This behaviour roughly matches both VS on Windows and VS Code